# persist: replace Abomonation with serde+bincode