--- a/src/persist/Cargo.toml
+++ b/src/persist/Cargo.toml
@@ -20,10 +20,9 @@ harness = false
 # NB: This is meant to be a strong, independent abstraction boundary, please
 # don't leak in deps on other Materialize packages.
 [dependencies]
-abomonation = "0.7"
-abomonation_derive = "0.5"
 async-trait = "0.1"
 aws-util = { path = "../aws-util" }
+bincode = "1.3.3"
 crossbeam-channel = "0.5"
 differential-dataflow = { git = "https://github.com/TimelyDataflow/differential-dataflow.git" }
 fail = { git = "https://github.com/tikv/fail-rs", features = ["failpoints"] }
@@ -34,6 +33,7 @@ ore = { path = "../ore", default-features = false, features = ["metrics"] }
 persist-types = { path = "../persist-types" }
 rusoto_core = "0.47.0"
 rusoto_s3 = "0.47.0"
+serde = { version = "1.0.130", features = ["derive"] }
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", default-features = false, features = ["bincode"] }
 tokio = { version = "1.13.0", default-features = false, features = ["macros", "sync", "rt"] }
 uuid = { version = "0.8.2", features = ["v4"] }
@@ -43,5 +43,4 @@ criterion = "0.3.5"
 ore = { path = "../ore", default-features = false, features = ["test"] }
 rand = { version = "0.8.4", features = [ "small_rng" ] }
 tempfile = "3.2.0"
-serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"