… Cargo.lock
… src/persist/src/golden_test.json