bincode isn't really a long term answer to our storage formats, but this
was easy and there were some concerns internally about using
Abomonation.

The performance regression is unfortunate, but not unexpected given how
little Abomonation actually had to do in the hot paths. We'll get all of
this back and more when we switch to using a columnar encoding for trace
and unsealed batches.

Before

    end_to_end/end_to_end_4b_keys
                            time:   [441.78 ms 443.93 ms 445.50 ms]
                            thrpt:  [22.447 MiB/s 22.526 MiB/s 22.635 MiB/s]
    end_to_end/end_to_end_4kb_keys
                            time:   [244.75 ms 245.27 ms 245.87 ms]
                            thrpt:  [40.671 MiB/s 40.772 MiB/s 40.859 MiB/s]
    end_to_end/end_to_end_100b_keys
                            time:   [244.18 ms 244.63 ms 245.06 ms]
                            thrpt:  [40.806 MiB/s 40.879 MiB/s 40.953 MiB/s]
    mem_unsealed_snapshot   time:   [28.079 ms 28.310 ms 28.533 ms]
    mem_trace_snapshot      time:   [28.738 ms 28.972 ms 29.189 ms]
    file_unsealed_snapshot  time:   [46.351 ms 46.783 ms 47.181 ms]
    file_trace_snapshot     time:   [47.693 ms 48.670 ms 49.672 ms]
    mem_write_sync          time:   [30.027 ns 30.182 ns 30.339 ns]
    file_write_sync         time:   [18.032 ms 18.212 ms 18.441 ms]
    indexed_write_drain     time:   [336.08 ms 337.22 ms 338.33 ms]
    indexed_write_drain_sorted
                            time:   [193.77 ms 194.83 ms 195.90 ms]

After

    end_to_end/end_to_end_4b_keys
                            time:   [443.18 ms 444.24 ms 445.30 ms]
                            thrpt:  [22.457 MiB/s 22.511 MiB/s 22.564 MiB/s]
                     change:
                            time:   [-1.2815% -0.4797% +0.2789%] (p = 0.28 > 0.05)
                            thrpt:  [-0.2782% +0.4820% +1.2981%]
                            No change in performance detected.
    end_to_end/end_to_end_4kb_keys
                            time:   [244.45 ms 245.04 ms 245.61 ms]
                            thrpt:  [40.716 MiB/s 40.810 MiB/s 40.909 MiB/s]
                     change:
                            time:   [-0.4210% +0.1634% +1.1124%] (p = 0.77 > 0.05)
                            thrpt:  [-1.1002% -0.1632% +0.4228%]
                            No change in performance detected.
    end_to_end/end_to_end_100b_keys
                            time:   [297.84 ms 302.01 ms 306.72 ms]
                            thrpt:  [32.603 MiB/s 33.112 MiB/s 33.575 MiB/s]
                     change:
                            time:   [+21.249% +22.825% +24.359%] (p = 0.00 < 0.05)
                            thrpt:  [-19.588% -18.583% -17.525%]
                            Performance has regressed.
    mem_unsealed_snapshot   time:   [28.075 ms 28.313 ms 28.542 ms]
                            change: [-1.0683% +0.0097% +1.0881%] (p = 0.99 > 0.05)
                            No change in performance detected.
    mem_trace_snapshot      time:   [28.583 ms 28.879 ms 29.184 ms]
                            change: [-1.6319% -0.3222% +0.9535%] (p = 0.63 > 0.05)
                            No change in performance detected.
    file_unsealed_snapshot  time:   [47.343 ms 47.626 ms 47.922 ms]
                            change: [+0.7467% +1.8011% +2.9477%] (p = 0.00 < 0.05)
                            Change within noise threshold.
    file_trace_snapshot     time:   [48.374 ms 48.756 ms 49.201 ms]
                            change: [-1.9911% +0.1774% +2.3542%] (p = 0.88 > 0.05)
                            No change in performance detected.
    mem_write_sync          time:   [30.055 ns 30.373 ns 30.847 ns]
                            change: [-13.544% +1.5000% +22.014%] (p = 0.87 > 0.05)
                            No change in performance detected.
    file_write_sync         time:   [18.055 ms 18.258 ms 18.509 ms]
                            change: [-1.4850% +0.2542% +1.9496%] (p = 0.78 > 0.05)
                            No change in performance detected.
    indexed_write_drain     time:   [341.91 ms 343.32 ms 344.82 ms]
                            change: [+1.2797% +1.8099% +2.3598%] (p = 0.00 < 0.05)
                            Performance has regressed.
    indexed_write_drain_sorted
                            time:   [202.51 ms 204.83 ms 207.35 ms]
                            change: [+3.8126% +5.1292% +6.6317%] (p = 0.00 < 0.05)
                            Performance has regressed.

<!--

Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste them
here, or write "see commit messages for details." If there is only one commit
in the PR, GitHub will have already added its commit message above.

-->

### Motivation

<!--

Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug. [Link to issue.]

  * This PR adds a known-desirable feature. [Link to issue.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

    [Write a brief specification for the feature, including justification
     for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

    [Describe what was wrong with the existing code, if it is not obvious.]

-->

### Tips for reviewer

<!--

Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
    * [Some function/module/file] is pure code movement and only needs a skim.

Delete this section if no tips.

-->

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly considered.
- [ ] This PR adds a release note for any
  [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/user/content/release-notes.md#what-changes-require-a-release-note).