@danhhz 2021-10-20
Note that the golden_test and the persist upgrade tests are expected to fail until we can bump the BlobMeta version.