@danhhz 2021-11-03
@ruchirK I went to update the benchmarks for this with your new ones but I'm getting an error that seems to reproduce at origin/main's HEAD.

```
$ git checkout 17b496784d6c9446588d56e5ea06e61122daf6b1
$ cargo bench -p persist -- blob_set/mem/7000000
...
     Running unittests (/tmp/target/release/deps/writer-9ad885949a465882)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
Benchmarking blob_set/mem/7000000: Collecting 100 samples in estimated 5.3386 s (1500 iterations)error: process didn't exit successfully: `/tmp/target/release/deps/writer-9ad885949a465882 blob_set/mem/7000000 --bench` (signal: 9, SIGKILL: kill)
```

Any ideas before I dig into this? I tried a `cargo test --benches -p persist -- blob_set/mem/7000000` and it seemed to hang