@ruchirK 2021-11-03
It's moving slowly because some of the benchmarks take 30 seconds but doing

```
cargo bench --benches writer
```

from within `src/persist` seems to work for me (at least `blob_set/mem/7000000` and a lot of the benches after it seem to work). When it finishes I'll try running from the repo toplevel