@ruchirK 2021-11-03
It seems to work from the repo toplevel as well for me:

```
altaria-2:materialize Test$ cargo bench -p persist -- blob_set/mem/7000000
warning: unused import: `HashSet`
  --> src/persist/src/indexed/mod.rs:22:43
   |
22 | use std::collections::{BTreeMap, HashMap, HashSet};
   |                                           ^^^^^^^
   |
   = note: `#[warn(unused_imports)]` on by default

   Compiling persist v0.0.0 (/Users/Test/github/materialize/src/persist)
warning: `persist` (lib) generated 1 warning
    Finished bench [optimized] target(s) in 35.00s
     Running unittests (target/release/deps/persist-b44f37356fbd715e)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 70 filtered out; finished in 0.00s

     Running unittests (target/release/deps/end_to_end-92b1b2f124a5def0)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend

     Running unittests (target/release/deps/snapshot-29a04297a206741b)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
     Running unittests (target/release/deps/writer-a6d4db66eafc8ffd)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
blob_set/mem/7000000    time:   [2.9954 ms 3.0282 ms 3.0665 ms]
                        thrpt:  [2.1260 GiB/s 2.1529 GiB/s 2.1765 GiB/s]
                 change:
                        time:   [-65.231% -57.458% -46.647%] (p = 0.00 < 0.05)
                        thrpt:  [+87.430% +135.06% +187.61%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

```