@ruchirK 2021-11-03
I have one observation and maybe one theory for why it crashed for you?

Observation: the tests seem kind of slow between benchmark groups - I _think_ that's because its spending time freeing memory and deleting the tempdir.

Theory: the in-memory blob / indexed tests are pretty ram hungry - it looked like the in-memory `set_unsealed_batch` benchmarks were using upwards for 15 gb for a bit --- maybe something is got oom killed? It's strange because the `blob_set` tests actually use a lot less ram than the others so it shouldn't happen when you only run that. Let me retry a few times to see if I can repro what you're getting.