@ruchirK 2021-11-03
@danhhz I reran `cargo bench -p persist -- blob_set/mem/7000000` several times -- it seems like its spiking up to 15 or gb of ram usage _after_ the text summary is printed to screen. I have no earthly idea what its spending that much ram on (this test only writes 7mb blobs), but after running it about 6 or 7 times I was unable to reproduce the crash you saw. Is it happening fairly regularly for you?