@danhhz 2021-11-03
Wow mine spiked up to 33GB of ram